+++ utils.h
@@ -5,7 +5,7 @@
 // Déclarations des fonctions utilitaires
 void initialize_random();
 float random_float(int min, int max);
-int estimate_max_laps(int session_duration, int max_time);
+int estimate_max_laps(int session_duration, float max_time);
 char *trim(char *str);
 
 #endif
